backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
timm